refactor: DRY up deser_nomad_number with macro
