+++ nomad-types/Cargo.toml
@@ -12,3 +12,4 @@ serde_json = { version = "1.0", default-features = false }
 thiserror = "*"
 tracing-subscriber = "0.2.15"
 prometheus = "0.12"
+paste = "1.0.5"
